# Speedup rust tests