Speeds up running the rust tests by about a factor of 2.

BTW using `cargo nextest` (after installing nextest e.g. with `pacman -Syu cargo-nextest` or `cargo install cargo-nextest`) you can see which tests take longest. 

With this PR, the longest tests are:
```
        PASS [   8.380s] deltachat context::tests::test_limit_search_msgs
        PASS [   6.775s] deltachat imex::tests::test_render_setup_file
        PASS [  11.286s] deltachat chat::tests::test_modify_chat_disordered
        PASS [   7.221s] deltachat imex::tests::test_render_setup_file_newline_replace
        PASS [  18.530s] deltachat blob::tests::test_recode_image_1
        PASS [  36.478s] deltachat blob::tests::test_recode_image_2
     Summary [  36.614s] 549 tests run: 549 passed, 0 skipped
```
(all other tests are under 3s) (note that tests are run in parallel)

Probably further optimizing isn't that important since compilation takes like 30s-1min.

#skip-changelog I guess, or do you think that's important enough?